Please remove the entire mission statement but keep the pictures for later use. John would also like the leadership pictures to be smaller, focusing more on name/title On the home page please remove the first sentence and keep "While most consulting firms conduct industry research and give advisory services, our team delivers solutions that can assist businesses and provide answers. We listen to management and proactively work together as a team."
We have a few corrections to make to the website as it stands.
starting with the knowledge center, it should read as follows;
"Our clients are mostly private businesses that want to expand, turn around their business, and or generate new investment ideas. Not every company is right for stock exchange, there's private ways to raise assets and to develop new ideas in this fast pace marketplace. We understand that not every client is the same and we ensure our strategy is right for particular clients' needs.
Our “Knowledge Center” consists of real knowledge that provides valuable data and expert opinions from industry specialists."